Split the crypto module into per-algorithm files
src/crypto.rs becomes a src/crypto/ module: mod.rs keeps the algorithm-agnostic
core (the Algorithm enum, the private SignatureScheme trait, scheme() dispatch,
sign_bytes/verify_bytes, the key types and key generation, and the shared
random_seed), while each algorithm's SignatureScheme implementation moves to its
own file — ed25519.rs (the only file that uses ed25519-dalek) and mldsa.rs (the
only file that uses the ml-dsa crate and getrandom). The dispatch calls
ed25519::scheme() / mldsa::scheme_44/65/87(); the trait and random_seed stay
private to crypto and the submodules reach them as descendants, so the public
surface is unchanged.

Confining each backend to one file makes swapping the ML-DSA implementation
(e.g. for a FIPS-validated module or an HSM behind the same SignatureScheme) a
single-file change.

No behaviour change: native + wasm builds, clippy, and rustfmt all clean; tests
pass unchanged.
